Deplace serializer infrastructure
