+++ geotrek/infrastructure/serializers.py
@@ -1,0 +1,35 @@
+from geotrek.authent.serializers import StructureSerializer
+from geotrek.common.serializers import PictogramSerializerMixin, BasePublishableSerializerMixin
+
+from geotrek.infrastructure import models as infrastructure_models
+
+
+class InfrastructureTypeSerializer(PictogramSerializerMixin):
+    class Meta:
+        model = infrastructure_models.InfrastructureType
+        fields = ('id', 'pictogram', 'label')
+
+
+class SignageSerializer(BasePublishableSerializerMixin):
+    type = InfrastructureTypeSerializer()
+    structure = StructureSerializer()
+
+    class Meta:
+        model = infrastructure_models.Signage
+        id_field = 'id'  # By default on this model it's topo_object = OneToOneField(parent_link=True)
+        geo_field = 'geom'
+        fields = ('id', 'structure', 'name', 'type') + \
+            BasePublishableSerializerMixin.Meta.fields
+
+
+class InfrastructureSerializer(BasePublishableSerializerMixin):
+    type = InfrastructureTypeSerializer()
+    structure = StructureSerializer()
+
+    class Meta:
+        model = infrastructure_models.Infrastructure
+        id_field = 'id'  # By default on this model it's topo_object = OneToOneField(parent_link=True)
+        geo_field = 'geom'
+        fields = ('id', ) + \
+            ('id', 'structure', 'name', 'type') + \
+            BasePublishableSerializerMixin.Meta.fields
