+++ geotrek/trekking/serializers.py
@@ -17,7 +17,7 @@ from geotrek.common.serializers import (
     PictogramSerializerMixin, ThemeSerializer,
     TranslatedModelSerializer, PicturesSerializerMixin,
     PublishableSerializerMixin, RecordSourceSerializer,
-    TargetPortalSerializer, BasePublishableSerializerMixin
+    TargetPortalSerializer
 )
 from geotrek.authent.serializers import StructureSerializer
 
@@ -25,7 +25,6 @@ from geotrek.cirkwi.models import CirkwiTag
 from geotrek.zoning.serializers import ZoningSerializerMixin
 from geotrek.altimetry.serializers import AltimetrySerializerMixin
 from geotrek.trekking import models as trekking_models
-from geotrek.infrastructure import models as infrastructure_models
 
 
 class TrekGPXSerializer(GPXSerializer):
@@ -318,37 +317,6 @@ class POISerializer(PublishableSerializerMixin, PicturesSerializerMixin,
             ZoningSerializerMixin.Meta.fields + \
             PublishableSerializerMixin.Meta.fields + \
             PicturesSerializerMixin.Meta.fields
-
-
-class InfrastructureTypeSerializer(PictogramSerializerMixin):
-    class Meta:
-        model = infrastructure_models.InfrastructureType
-        fields = ('id', 'pictogram', 'label')
-
-
-class SignageSerializer(BasePublishableSerializerMixin):
-    type = InfrastructureTypeSerializer()
-    structure = StructureSerializer()
-
-    class Meta:
-        model = infrastructure_models.Signage
-        id_field = 'id'  # By default on this model it's topo_object = OneToOneField(parent_link=True)
-        geo_field = 'geom'
-        fields = ('id', 'structure', 'name', 'type') + \
-            BasePublishableSerializerMixin.Meta.fields
-
-
-class InfrastructureSerializer(BasePublishableSerializerMixin):
-    type = InfrastructureTypeSerializer()
-    structure = StructureSerializer()
-
-    class Meta:
-        model = infrastructure_models.Infrastructure
-        id_field = 'id'  # By default on this model it's topo_object = OneToOneField(parent_link=True)
-        geo_field = 'geom'
-        fields = ('id', ) + \
-            ('id', 'structure', 'name', 'type') + \
-            BasePublishableSerializerMixin.Meta.fields
 
 
 class ServiceTypeSerializer(PictogramSerializerMixin, TranslatedModelSerializer):
